FloatArray debug variable type for calibration (#1164)
* "trace implementation cleanup; updates #1154"

* "variable length encoded, but only used trivially, errors in debug protocol; updates #1154"

* "trace handler and circular buffer errors fixed; updates #1154"

* "disabled obsolete trace handler test; updates #1154"

* "FloatArray debug var type and tests; updates #1154"

* "float array implemented and debug interface reads values; updates #1154"

* "FloatArray can be manipulated via debug interface; updates #1154"

* "each pinch valve keeps its own calibration, initial values hard-coded; updates #1154"

* "fixed integration test bug, improved FloatArray test; updates #1154"

* "removing redundant gitignore; updates #1154"

* "rule to ignore rendudant gitignore; updates #1154"

* "refactoring, documentation and more tests for Trace and its command handler, fixes in python interface; updates #1154"

* "more cleanup and warnings; updates #1154"

* "reverting overzealous replace; updates #1154"

* "addressed PR comments; updates #1154"

+++ software/controller/lib/hal/i2c.h
@@ -96,11 +96,11 @@ class Channel {
  protected:
   // We queue of a few requests. The number of requests is arbitrary but
   // should be enough for all intents and purposes.
-  static constexpr uint8_t QueueLength{80};
+  static constexpr size_t QueueLength{80};
 
   // We copy the write data into a buffer to make sure nothing can be lost
   // due to the scope of the caller's variable. This is the buffer size.
-  static constexpr uint32_t WriteBufferSize{4096};
+  static constexpr size_t WriteBufferSize{4096};
 
   // Max retry-after-error allowed for a single request.
   static constexpr int8_t MaxRetries{5};
@@ -171,9 +171,9 @@ class Channel {
   //    means we lose the ability to retry a request after an I²C (or DMA)
   //    error.
   uint8_t write_buffer_[WriteBufferSize];
-  uint32_t write_buffer_index_{0};
-  uint32_t write_buffer_start_{0};
-  uint32_t wrapping_index_{WriteBufferSize};
+  size_t write_buffer_index_{0};
+  size_t write_buffer_start_{0};
+  size_t wrapping_index_{WriteBufferSize};
   bool CopyDataToWriteBuffer(const void *data, uint16_t size);
 };
